Refactor: MessageHandler type 파일 추가
